lint

+++ libs/blockmesh-bin/src-tauri/src/ore/ore_process_wrapper.rs
@@ -1,7 +1,9 @@
+use anyhow::anyhow;
 use std::env;
 use std::fs::File;
 use std::path::PathBuf;
 use std::process::Stdio;
+use std::sync::Arc;
 use sysinfo::{Pid, System};
 use tokio::process::Command;
 use tokio::sync::Mutex;
@@ -9,6 +11,7 @@ use tokio::task::JoinHandle;
 
 use block_mesh_common::app_config::TaskStatus;
 
+use crate::tauri_state::AppState;
 use crate::SYSTEM;
 
 #[derive(Debug)]
@@ -18,7 +21,6 @@ pub struct OreProcessInput {
     pub ore_keypair: String,
     pub ore_priority_fee: u64,
     pub ore_rpc: String,
-    pub ore_tx: tokio::sync::mpsc::Sender<TaskStatus>,
 }
 
 #[derive(Debug)]
@@ -29,8 +31,25 @@ pub struct OreProcessOutput {
     pub stderr_log: String,
 }
 
-#[tracing::instrument(name = "ore_process_wrapper", skip(params), ret, err)]
-pub async fn ore_process_wrapper(params: OreProcessInput) -> anyhow::Result<OreProcessOutput> {
+#[tracing::instrument(name = "ore_process_wrapper", skip(app_state), ret, err)]
+pub async fn start_ore_process(
+    app_state: Arc<Mutex<AppState>>,
+) -> anyhow::Result<OreProcessOutput> {
+    let mut state = app_state.lock().await;
+
+    let exec = std::env::current_exe().unwrap_or_default();
+    let ore = exec.parent().unwrap().join("resources").join("ore");
+    if tokio::fs::metadata(&ore).await.is_err() {
+        log::error!("ORE binary not found at {:?}", ore);
+        state.config.ore_status = Option::from(TaskStatus::Off);
+        return Err(anyhow!("ORE binary not found"));
+    }
+    let ore_exec = ore.to_str().unwrap_or_default().to_string();
+    let ore_threads = state.config.ore_threads.unwrap_or_default();
+    let ore_keypair = state.config.ore_keypair.clone().unwrap_or_default();
+    let ore_priority_fee = state.config.ore_priority_fee.unwrap_or_default();
+    let ore_rpc = state.config.ore_rpc.clone().unwrap_or_default();
+
     let temp = env::temp_dir();
     let stdout_log_file = PathBuf::new().join(temp.clone()).join("ore_stdout.log");
     if tokio::fs::metadata(&stdout_log_file).await.is_err() {
@@ -44,16 +63,16 @@ pub async fn ore_process_wrapper(params: OreProcessInput) -> anyhow::Result<OreP
     let stderr_log = stderr_log_file.to_str().unwrap().to_string();
     let stdout_log_file = File::open(&stdout_log_file)?;
     let stderr_log_file = File::open(&stderr_log_file)?;
-    let mut child = Command::new(&params.ore_exec)
+    let mut child = Command::new(&ore_exec)
         .arg("mine")
         .arg("--rpc")
-        .arg(params.ore_rpc)
+        .arg(ore_rpc)
         .arg("--threads")
-        .arg(params.ore_threads.to_string())
+        .arg(ore_threads.to_string())
         .arg("--keypair")
-        .arg(params.ore_keypair)
+        .arg(ore_keypair)
         .arg("--priority-fee")
-        .arg(params.ore_priority_fee.to_string())
+        .arg(ore_priority_fee.to_string())
         .stdout(Stdio::from(stdout_log_file))
         .stderr(Stdio::from(stderr_log_file))
         .spawn()?;
@@ -61,7 +80,6 @@ pub async fn ore_process_wrapper(params: OreProcessInput) -> anyhow::Result<OreP
     let pid = child.id().ok_or(anyhow::anyhow!("Failed to get PID"))?;
     let handle: JoinHandle<()> = tokio::spawn(async move {
         let _ = child.wait().await;
-        params.ore_tx.send(TaskStatus::Off).await.unwrap();
     });
 
     Ok(OreProcessOutput {
@@ -72,16 +90,23 @@ pub async fn ore_process_wrapper(params: OreProcessInput) -> anyhow::Result<OreP
     })
 }
 
-#[tracing::instrument(name = "kill_ore_process", ret, err)]
-pub async fn kill_ore_process(pid: u32) -> anyhow::Result<()> {
-    let system = SYSTEM.get_or_init(|| Mutex::new(System::new_all()));
-    let mut system = system.lock().await;
-    system.refresh_process(Pid::from_u32(pid));
-    system
-        .processes_by_name("ore")
-        .filter(|process| process.pid().as_u32() == pid)
-        .for_each(|process| {
-            process.kill();
-        });
+#[tracing::instrument(name = "kill_ore_process", skip(app_state), ret, err)]
+pub async fn kill_ore_process(app_state: Arc<Mutex<AppState>>) -> anyhow::Result<()> {
+    tracing::info!("here 95");
+    let mut state = app_state.lock().await;
+    tracing::info!("here 97");
+    if let Some(pid) = state.ore_pid {
+        let system = SYSTEM.get_or_init(|| Mutex::new(System::new_all()));
+        let mut system = system.lock().await;
+        system.refresh_process(Pid::from_u32(pid));
+        system
+            .processes_by_name("ore")
+            .filter(|process| process.pid().as_u32() == pid)
+            .for_each(|process| {
+                process.kill();
+            });
+        state.config.ore_status = Option::from(TaskStatus::Off);
+        state.ore_pid = None;
+    }
     Ok(())
 }
